% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_build.R
\name{bootstrap.build}
\alias{bootstrap.build}
\title{Perform network inference with bootstrap approach.}
\usage{
bootstrap.build(expr, nboot = 10, nkeep1 = round(0.1 * nrow(expr)^2/2),
  nkeep2 = round(0.5 * nboot), parallel = FALSE, ncores = 10,
  seeds = NULL, ...)
}
\arguments{
\item{expr}{a n by d data matrix}

\item{nboot}{number of bootstrap samples}

\item{nkeep1}{number of edges to keep from each infered network}

\item{nkeep2}{minimum number of times an edge has to be predicted to be 
kept in the final network}

\item{parallel}{should the function be run in parallel (unix-like systems
only)}

\item{ncores}{number of cores for parallel executions}

\item{seeds}{vector of \code{nboot} seeds to control randomness in network 
inference based on bootstraps}

\item{...}{options passed to the function \code{\link{set.options}}}
}
\value{
object with the following entries
\itemize{
\item{\code{edges} final infered network}
\item{\code{mse} total OOB mean square error}
\item{\code{adjacency} total adjacency matrix combined from the \code{nboot}
infered network}
\item{\code{stars} StaRs criterion}
\item{\code{stability} stability criterion}
}
}
\description{
Perform network inference with bootstrap approach.
}
\examples{
data(cancer)
# no constraint
res <- bootstrap.build(expr, nboot=3, nkeep1=10, nkeep2=1)
# constraints
res <- bootstrap.build(expr, nboot=3, nkeep1=10, nkeep2=1,
                       first.list=matrix(c("AMFR","BB_S4","BECNI","BTG3"),
                                         ncol=2, byrow=TRUE),
                       second.list=matrix(c("AMFR","E2F3"), ncol=2),
                       mu=1)
print(res)
}
