% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_network.R
\name{build.network}
\alias{build.network}
\alias{print.internet3D}
\alias{summary.internet3D}
\title{Perform network inference.}
\usage{
build.network(expr, ...)

\method{print}{internet3D}(x, ...)

\method{summary}{internet3D}(object, ...)
}
\arguments{
\item{expr}{a n by d data matrix}

\item{...}{options passed to the function \code{\link{set.options}}}

\item{x}{a \code{internet3D} object}

\item{object}{a \code{internet3D} object}
}
\value{
object of type \code{internet3D} with the following entries 
\itemize{
\item{\code{expr} input expression dataset}
\item{\code{networks} networks infered along the regularization path}
\item{\code{used.lambdas} penalization parameters used along the 
regularization path}
\item{\code{n.edges} number of edges of networks along the regularization 
path}
\item{\code{BIC} BIC criteria (experimental) along the regularization path}
\item{\code{AIC} AIC criteria (experimental) along the regularization path}
\item{\code{loglik} log-likelihoods along the regularization path}
\item{\code{loglik.pen} penalized log-likelihoods along the regularization 
path}
\item{\code{options} \code{internet3DOptions} object with the options used 
for the inference}
}
}
\description{
Perform network inference.
}
\examples{
data(cancer)
# no constraint
res <- build.network(expr)
# constraints
res <- build.network(expr, 
                     first.list=matrix(c("AMFR","BTG3","BECNI","BTG3"),
                                       ncol=2, byrow=TRUE),
                     second.list=matrix(c("AMFR","E2F3"), ncol=2),
                     mu=1)
print(res)

}
\seealso{
\code{\link{set.options}} and \code{\link{bootstrap.build}}
}
