% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_options.R
\name{set.options}
\alias{set.options}
\alias{print.internet3DOptions}
\title{Set and print options for network inference.}
\usage{
set.options(first.list = NULL, second.list = NULL, scale = TRUE,
  verbose = TRUE, mu = NULL, penalties = NULL, penalty.min = 0.01,
  penalty.max = NULL, n.penalties = 100, edges.max = Inf,
  symmetrization = c("AND", "OR"), initial.guess = NULL, max.it = NULL)

\method{print}{internet3DOptions}(x, ...)
}
\arguments{
\item{first.list}{list of edges supposed to be present (array with two 
columns and a number of rows corresponding to the number of edges with a 
prior)}

\item{second.list}{list of edges supposed to be absent (array with two 
columns and a number of rows corresponding to the number of edges with a 
prior)}

\item{scale}{logical. Should the data be scaled to unit variance prior the 
analysis}

\item{verbose}{logical. Should messages be printed during the learning
process}

\item{mu}{positive number. Regularization parameter for the L2 penalty on
edges with a prior assumption}

\item{penalties}{a sequence of decreasing positive numbers to control the L1
regularization}

\item{penalty.min}{if \code{penalties=NULL}, minimum value for the L1
regularization parameter}

\item{penalty.max}{if \code{penalties=NULL}, maximum value for the L1
regularization parameter}

\item{n.penalties}{if \code{penalties=NULL}, number of values to include in
the sequence of positive numbers that control the L1 regularization}

\item{edges.max}{maximum number of edges to stop the learning process}

\item{symmetrization}{symmetrization rule, to be chosen between \code{"AND"}
and \code{"OR"}}

\item{initial.guess}{initial guess (if existing) for the solution of the
optimization problem}

\item{max.it}{maximum number of iterations allowed to solve the optimization 
problem}

\item{x}{a \code{internet3DOptions} object}

\item{...}{not used}
}
\value{
object of type \code{internet3DOptions} that can be used within the
functions \code{\link{build.network}} or \code{\link{bootstrap.build}}
}
\description{
Set and print options for network inference.
}
\examples{
set.options()
set.options(mu = 1)
}
\seealso{
\code{\link{build.network}}, \code{\link{bootstrap.build}}
}
